# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Book Application API
  description: Sample Spring Boot REST API application to manage books using springdoc-openapi
    and OpenAPI 3.
  contact:
    email: towerspro@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/users/{username}:
    get:
      tags:
      - user-controller
      summary: Get a user by its id
      operationId: getUserById
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: Found the user
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        "404":
          description: user not found
    put:
      tags:
      - user-controller
      summary: Updates a user
      operationId: updateUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: If the user has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        "500":
          description: If a runtime error occurs while processing the request
        "400":
          description: If one or more request parameters don't comply with the specification
    delete:
      tags:
      - user-controller
      summary: Delete a user by its id
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: If the user has been deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        "400":
          description: Invalid id supplied
        "404":
          description: user not found
  /api/v1/users/:
    get:
      tags:
      - user-controller
      summary: Get a list of users
      operationId: getUsers
      responses:
        "200":
          description: Returns the list of users
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        "500":
          description: If a runtime error occurs while processing the request
    post:
      tags:
      - user-controller
      summary: Creates a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: If the user has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        "500":
          description: If a runtime error occurs while processing the request
        "400":
          description: If one or more request parameters don't comply with the specification
  /api/v1/books/{id}/comments:
    post:
      tags:
      - book-controller
      summary: Creates a new comment
      operationId: createComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        "200":
          description: If the comment has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        "500":
          description: If a runtime error occurs while processing the request
        "400":
          description: If one or more request parameters don't comply with the specification
  /api/v1/books/:
    get:
      tags:
      - book-controller
      summary: Get a list of books
      operationId: getBooks
      responses:
        "200":
          description: Returns the list of books
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookListResponseDTO'
        "500":
          description: If a runtime error occurs while processing the request
    post:
      tags:
      - book-controller
      summary: Creates a new book
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookPostDTO'
        required: true
      responses:
        "200":
          description: If the book has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        "500":
          description: If a runtime error occurs while processing the request
        "400":
          description: If one or more request parameters don't comply with the specification
  /api/v1/users/{userId}/comments:
    get:
      tags:
      - user-controller
      summary: Get all the comments from a user
      operationId: getCommentsByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: Found the comments
          content:
            '*/*':
              schema:
                uniqueItems: true
                type: array
                items:
                  $ref: '#/components/schemas/CommentDTO'
        "404":
          description: Comment not found
  /api/v1/books/{id}:
    get:
      tags:
      - book-controller
      summary: Get a book by its id
      operationId: getBookById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Book not found
        "400":
          description: Invalid id supplied
        "200":
          description: Found the book
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      tags:
      - book-controller
      summary: Delete a book by its id
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Book not found
        "400":
          description: Invalid id supplied
        "204":
          description: If the book has been deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /api/v1/books/{bookId}/comments/{id}:
    get:
      tags:
      - book-controller
      summary: Get a comment by its id
      operationId: getCommentByBookAndId
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: Found the comment
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Comment'
        "404":
          description: Comment not found
    delete:
      tags:
      - book-controller
      summary: Delete a comment by its id
      operationId: deleteComment
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "204":
          description: If the comment has been deleted successfully
        "404":
          description: Comment not found
components:
  schemas:
    UserDTO:
      type: object
      properties:
        email:
          type: string
        nickname:
          type: string
    Comment:
      required:
      - text
      - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          maxLength: 20
          minLength: 2
          type: string
        text:
          maxLength: 200
          minLength: 3
          type: string
        rating:
          maximum: 5
          minimum: 0
          type: integer
          format: int32
    BookPostDTO:
      required:
      - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          maxLength: 200
          minLength: 2
          type: string
        summary:
          type: string
        author:
          type: string
        publisher:
          type: string
        publicationYear:
          type: integer
          format: int32
    Book:
      required:
      - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          maxLength: 200
          minLength: 2
          type: string
        summary:
          type: string
        author:
          type: string
        publisher:
          type: string
        publicationYear:
          type: integer
          format: int32
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponseDTO'
    CommentResponseDTO:
      type: object
      properties:
        text:
          type: string
        nick:
          type: string
        email:
          type: string
    CommentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        text:
          type: string
        rating:
          type: integer
          format: int32
        book_id:
          type: integer
          format: int64
    BookListResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string